Parent inverse display: proper scale manipulation, remove scene handler
Handler was using CPU at all times, so create button to fetch inverse on demand.

Fixed scale manipulation forcing to 1.0. Scale matrices are created one axis at a time then multiplied to achieve overall scale matrix.

TODO:
Operator to clear parent inverse and apply transforms to the object. Maybe not a good idea because of possibility to collide with animation.